xsm: Don't audit argo enable xsm check during initialization
argo_init calls xsm_argo_enable to see if the policy allows access
to argo.  If it does, then argo is setup for the domain - otherwise it
is not initialized.

When a Flask policy selectively enables argo for only certain domain
types, creating domains without the permission generates an AVC.
Generating an AVC is misleading since it's Xen calling XSM to make an
initialization decision and not the domain trying to access argo.

Add xsm_flask_enable_noaudit to bypass the AVC generation for this case.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>