Importing patch config.patch
SHORT DESCRIPTION:
Xen tools configuration script fixes.

LONG DESCRIPTION:
This patch removes a number of things that we do not want configure to check
for in our environment. This should be able to be done with arguments to the
script but earlier versions of it did not have the means to do that.

CHANGELOG

REMOVAL
It may be that these things can be turned off with arguments to the script. If
so then this patch can go.

UPSTREAM PLAN
No.

INTERNAL DEPENDENCIES

PATCHES
--- a/tools/configure