Importing patch opt-disable-vmcs-shadowing.patch
################################################################################
SHORT DESCRIPTION:
################################################################################
Introduced option to disable VMCS shadowing.

################################################################################
LONG DESCRIPTION:
################################################################################
This patch was created to support vSentry running as a nested VMM on OpenXT.
It is not entirely clear why VMCS shadowing needs to be disabled since it is a
feature to assist nested VMMs. It may not be needed any longer to support
vSentry but it will be left here as a placeholder since it does nothing by
default.

################################################################################
CHANGELOG
################################################################################

################################################################################
REMOVAL
################################################################################
If no longer needed for vSentry interop, then it can be removed.

################################################################################
UPSTREAM PLAN
################################################################################

################################################################################
INTERNAL DEPENDENCIES
################################################################################

################################################################################
PATCHES
################################################################################
--- a/xen/arch/x86/hvm/vmx/vmcs.c