Importing patch libxl-disable-json-updates.patch
################################################################################
SHORT DESCRIPTION:
################################################################################
Disable libxl's functionality for rewriting VM config files when changes are
applied to the running system because they are not there and it causes failure.

################################################################################
LONG DESCRIPTION:
################################################################################
In Xen 4.9, libxl rewrites a VM config file when its running configuration
changes, eg. when adding a network interface to a VM.

OpenXT's toolstack does not currently use or provide the VM config files that
libxl expects so the attempted config rewrite fails, which disables all guest
networking.

Upstream Xen does not provide configuration or command line options to disable
the behaviour, so this patch disables it in the source.

################################################################################
CHANGELOG
################################################################################
Authors:
Christopher Clark <christopher.clark6@baesystems.com>

Introduced in the OpenXT uprev from Xen 4.6 to 4.9.

################################################################################
REMOVAL
################################################################################
OpenXT could integrate further with libxl, making changes to use the config
files that libxl expects. That would allow this patch to be removed.

Alternatively, patches could be submitted upstream to make the use of libxl's
config files optional and OpenXT could continue to avoid it.

################################################################################
UPSTREAM PLAN
################################################################################
Not as implemented: upstream evidently has a requirement for this feature.

################################################################################
INTERNAL DEPENDENCIES
################################################################################

################################################################################
PATCHES
################################################################################
--- a/tools/libxl/libxl_internal.h