Importing patch libxl-vwif-support.patch
################################################################################
SHORT DESCRIPTION:
################################################################################
libxl vwif support (deprecated)

################################################################################
LONG DESCRIPTION:
################################################################################
This patch overloads NICs with a second "kind": VWIF.
VWIF NICs are wireless NICs, which can be driven in the backend by netback (plus
some helper to fill in the fake wireless access point info in xenstore).
A custom frontend (like xenwnet) is required to expose a wireless nic to the
guest.

As of Xen 4.11, I consider this patch deprecated.
libxl (over-)uses macros and generic functions to handle devices, and those are
not compatible with "multi-kind" device types.
To deal with that, this patch keeps introducing more and more really ugly hacks.
A more correct way to do this would be to introduce a new type of device
alltogether. But that would lead to a massive amount of code duplication,
as "wnics" would be parsed and handled almost exactly like nics.

The main motivation behind VWIFs is to show the wireless signal strength inside
the guest, allowing for a better user experience.
But just for that small tray icon, a whole driver has to be maintained for all
supported guest operating systems, as well as some backend code.
See the REMOVAL section below for a simpler solution.

################################################################################
CHANGELOG
################################################################################
Rebased for Xen 4.8, 4.9, 4.11

################################################################################
REMOVAL
################################################################################
A small app should be written for every supported operating system.
That app would have a tray icon that shows the signal strength, while allowing
the NIC to be a regular VIF, driven by the standard netfront.
With such an app, this patch could be discarded, and frontend driver(s) could be
thrown away.

See https://openxt.atlassian.net/browse/OXT-1455

################################################################################
UPSTREAM PLAN
################################################################################
Please don't.

################################################################################
INTERNAL DEPENDENCIES
################################################################################
Some toolstack code to write vlan nodes and custom frontend drivers like xenwnet

################################################################################
PATCHES
################################################################################
--- a/tools/libxl/libxl_device.c