Importing patch argo-fix-full-ring-write-bug.patch
################################################################################
SHORT DESCRIPTION:
################################################################################
Fix issue of argo messages not being resent if ring is full.

################################################################################
LONG DESCRIPTION:
################################################################################
The argo kernel module depends on Xen's sendv() returning -EAGAIN if Xen
tries to write to a ring that is too full. Xen, however, will instead return
the value of pending_requeue (the requeue mechanism), which will always
resolve to 0 on success. This gives the caller no way of knowing whether or
not it should try sending again at a later time.

Now, if the ring is too full, Xen will only overwrite the return code with
that of pending_requeue() if pending_requeue returns a non-zero exit code.

################################################################################
CHANGELOG:
################################################################################

################################################################################
REMOVAL:
################################################################################
This can be removed when the bug is fixed upstream.

################################################################################
UPSTREAM PLAN:
################################################################################
Patch has been submitted upstream and is pending approval.

################################################################################
INTERNAL DEPENDENCIES:
################################################################################

################################################################################
PATCHES:
################################################################################
--- a/xen/common/argo.c