Importing patch libxl-fix-flr.patch
################################################################################
SHORT DESCRIPTION:
################################################################################
Fix device reset for pci devices in libxl

################################################################################
LONG DESCRIPTION:
################################################################################
This patch fixes the logic and the control flow path for pci device reset in
libxl.  Additionally, it addresses an issue with Xenbus state synchronization
during domain teardown.  More detailed information about the patch follows:

The changes to Xenbus state are to prevent hitting a timeout when domains are
shutdown or destroyed. Xl is waiting for "XenbusStateConnected" (state 4), while
the state machine has already moved the device to Closing (state 5) and Closed
(state 6) by the time we reach this point tearing down the domain.

The reset logic introduced in this patch is delayed until the domain has fully
been destroyed. The reason for this has to do with the
thorough-reset-interface-to-pciback-s-sysfs.patch in the linux patchqueue.
__pcistub_raw_device_reset tries multiple approaches for resetting the device,
flr, slot-level, and bus-level. If we encounter a device that, for example,
doesn't support flr but does support a slot-level reset, we want to make sure
all the functions on that device (GPUS are a good example, they often have 2
functions, video and HDMI audio, and are the only device in the slot) are
released from the domain before attempting the reset or it will fail. This
approach seems to be taken in an attempt to support a wide variety of PCI
devices.

################################################################################
CHANGELOG
################################################################################
Authors:
Chris Rogers <rogersc@ainfosec.com>

################################################################################
REMOVAL
################################################################################
N/A

################################################################################
UPSTREAM PLAN
################################################################################
Depends on the feasibility of upstreaming thorough-reset-interface-to-pciback...

################################################################################
INTERNAL DEPENDENCIES
################################################################################

################################################################################
PATCHES
################################################################################
--- a/tools/libxl/libxl_internal.h