Importing patch argo-fix-requeue-msg-len-bug.patch
################################################################################
SHORT DESCRIPTION:
################################################################################
Fix issue of Xen spurriously waking domains to write more data
even if receive ring is full.

################################################################################
LONG DESCRIPTION:
################################################################################
When a message is requeue'd in Xen's internal queue, the queue entry
contains the length of the message so that Xen knows to send a VIRQ to
the respective domain when enough space frees up in the ring. Due to a
small bug, however, Xen doesn't populate the length of the msg if a given
write fails, so this length is always reported as zero. This causes Xen
to spurriously wake up a domain even when the ring is still full.

This patch makes sure that the msg len is properly reported by populating
it in the event of a write failure.

################################################################################
CHANGELOG:
################################################################################

################################################################################
REMOVAL:
################################################################################
This can be removed when the bug is fixed upstream.

################################################################################
UPSTREAM PLAN:
################################################################################
Patch has been submitted upstream and is pending approval.

################################################################################
INTERNAL DEPENDENCIES:
################################################################################

################################################################################
PATCHES:
################################################################################
--- a/xen/common/argo.c